refactor(SPV-1087): extract struct types into separate files.
